+++ wellregistry/registry/models.py
@@ -3,6 +3,7 @@ Well Registry ORM object.
 """
 
 from django.conf import settings
+from django.core.exceptions import ValidationError
 from django.core.validators import RegexValidator
 from django.db import models
 
@@ -140,8 +141,8 @@ class MonitoringLocation(models.Model):
 
     agency = models.ForeignKey(AgencyLookup, on_delete=models.PROTECT, db_column='agency_cd', null=True,
                                to_field='agency_cd')
-    site_no = models.CharField(max_length=16)
-    site_name = models.CharField(max_length=300)
+    site_no = models.CharField(max_length=16, validators=[non_blank_validator])
+    site_name = models.CharField(max_length=300, validators=[non_blank_validator])
 
     country = models.ForeignKey(CountryLookup, on_delete=models.PROTECT, db_column='country_cd',
                                 null=True, blank=True, to_field='country_cd')
@@ -181,7 +182,7 @@ class MonitoringLocation(models.Model):
     alt_method = models.CharField(max_length=300, blank=True, verbose_name='Altitude method')
     alt_acy = models.CharField(max_length=300, blank=True, verbose_name='Altitude accuracy')
 
-    well_depth = models.DecimalField(max_digits=11, decimal_places=7, null=True)
+    well_depth = models.DecimalField(max_digits=11, decimal_places=7, null=True, blank=True)
     well_depth_units = models.ForeignKey(UnitsLookup, related_name='+', db_column='well_depth_units',
                                          on_delete=models.PROTECT, to_field='unit_id', null=True, blank=True)
 
@@ -231,6 +232,42 @@ class MonitoringLocation(models.Model):
 
     class Meta:
         unique_together = (('site_no', 'agency'),)
+
+    def clean(self):
+        """
+        Override model clean to do multi field validation
+        """
+        if self.site_type == 'WELL' and self.aqfr_type == '':
+            raise ValidationError(
+                'If the site is of type "WELL", then you must enter an Aquifer type')
+
+        if self.well_depth and not self.well_depth_units:
+            raise ValidationError(
+                'If Well depth is populated, then you must enter a Well depth unit')
+
+        if not self.well_depth and self.well_depth_units:
+            raise ValidationError(
+                'If Well depth is not populated, then Well depth unit must be left blank')
+
+        if (self.display_flag and self.wl_sn_flag) and \
+                (self.wl_well_type == '' or self.wl_well_purpose == ''):
+            raise ValidationError(
+                'If the well is In WL sub-network, then you must enter a WL well type and WL well purpose')
+
+        if (self.display_flag and self.wl_sn_flag and
+                self.wl_baseline_flag and self.wl_well_chars == ''):
+            raise ValidationError(
+                'If the well is in WL sub-network and in WL Baseline, then you must enter WL Well Characteristics')
+
+        if (self.display_flag and self.qw_sn_flag) and \
+                (self.qw_well_type == '' or self.qw_well_purpose == ''):
+            raise ValidationError(
+                'If the well is In QW sub-network, then you must enter a QW well type and QW well purpose')
+
+        if (self.display_flag and self.qw_sn_flag and
+                self.qw_baseline_flag and self.qw_well_chars == ''):
+            raise ValidationError(
+                'If the well is in QW sub-network and in WL Baseline, then you must enter QW Well Characteristics')
 
     def __str__(self):
         """Default string."""